requirements.txt
torch
torchvision
opencv-python
pillow
numpy
pytesseract
streamlit
